requirements.txt
fastapi
uvicorn
streamlit
websocket-client
python-dotenv
langchain
google-generativeai
langchain-groq
langchain-google-genai